refactor: contracts package cleanup
moved DU2 contracts to du2 subfolders, du3 stuff to main folder, eliminated all mentions of "unichain". It's DU3 now.
