+++ packages/contracts/contracts/du2/DataUnionMainnet.sol
@@ -4,12 +4,12 @@ pragma solidity 0.8.6;
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
  // TODO: switch to "@openzeppelin/contracts/access/Ownable.sol";
-import "./Ownable.sol";
-import "./PurchaseListener.sol";
+import "../Ownable.sol";
+import "../PurchaseListener.sol";
 import "./CloneLib.sol";
 import "./xdai-mainnet-bridge/IAMB.sol";
 import "./xdai-mainnet-bridge/ITokenMediator.sol";
-import "./IERC677Receiver.sol";
+import "../IERC677Receiver.sol";
 
 contract DataUnionMainnet is Ownable, PurchaseListener, IERC677Receiver {
