+++ packages/contracts/index.js
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-const { abi: templateAbi, bytecode: templateBytecode } = require("./artifacts/contracts/unichain/DataUnionTemplate.sol/DataUnionTemplate.json")
-const { abi: factoryAbi, bytecode: factoryBytecode } = require("./artifacts/contracts/unichain/DataUnionFactory.sol/DataUnionFactory.json")
+const { abi: templateAbi, bytecode: templateBytecode } = require("./artifacts/contracts/DataUnionTemplate.sol/DataUnionTemplate.json")
+const { abi: factoryAbi, bytecode: factoryBytecode } = require("./artifacts/contracts/DataUnionFactory.sol/DataUnionFactory.json")
 const { abi: oracleAbi, bytecode: oracleBytecode } = require("./artifacts/contracts/DefaultFeeOracle.sol/DefaultFeeOracle.json")
 
 module.exports = {
