+++ packages/contracts/scripts/du2/deploy-du-template-mainnet.ts
