+++ packages/contracts/scripts/du2/deploy-mainnet-xdai.ts
