+++ packages/contracts/test/contracts/DataUnionTemplate.test.ts
@@ -3,16 +3,14 @@ import { waffle } from "hardhat"
 import { BigNumber, Wallet, Contract, utils, BigNumberish } from "ethers"
 
 import Debug from "debug"
-const log = Debug("Streamr:du:test:DataUnionSidechain")
+const log = Debug("Streamr:du:test:DataUnionTemplate")
 //const log = console.log  // for debugging?
 
-import DataUnionSidechainJson from "../../artifacts/contracts/DataUnionSidechain.sol/DataUnionSidechain.json"
-
+import DataUnionTemplateJson from "../../artifacts/contracts/DataUnionTemplate.sol/DataUnionTemplate.json"
 import TestTokenJson from "../../artifacts/contracts/test/TestToken.sol/TestToken.json"
-import MockTokenMediatorJson from "../../artifacts/contracts/test/MockTokenMediator.sol/MockTokenMediator.json"
-import MockAMBJson from "../../artifacts/contracts/test/MockAMB.sol/MockAMB.json"
+import FeeOracleJson from "../../artifacts/contracts/DefaultFeeOracle.sol/DefaultFeeOracle.json"
 
-import type { DataUnionSidechain, MockTokenMediator, TestToken, MockAMB } from "../../typechain"
+import type { DataUnionTemplate, TestToken, DefaultFeeOracle } from "../../typechain"
 
 type EthereumAddress = string
 
@@ -31,7 +29,7 @@ const { hexZeroPad, parseEther, arrayify } = utils
  * @param {EthereumAddress} signer who authorizes withdraw
  * @param {EthereumAddress} to who gets the tokens
  * @param {number} amountTokenWei tokens multiplied by 10^18, or zero for unlimited (withdrawAllToSigned)
- * @param {Contract} duContract DataUnionSidechain contract object
+ * @param {Contract} duContract DataUnionTemplate contract object
  * @param {number} previouslyWithdrawn (optional) amount of token-wei withdrawn at the moment this signature is used
  */
 async function getWithdrawSignature(
@@ -48,10 +46,10 @@ async function getWithdrawSignature(
     return signer.signMessage(arrayify(message))
 }
 
-describe("DataUnionSidechain", () => {
+describe("DataUnionTemplate", () => {
     const accounts = provider.getWallets()
-    const creator = accounts[0]
-    const duBeneficiary = accounts[1]
+    const dao = accounts[0]
+    const admin = accounts[1]
     const agents = accounts.slice(2, 4)
     const members = accounts.slice(4, 7)
     const others = accounts.slice(7)
@@ -61,61 +59,52 @@ describe("DataUnionSidechain", () => {
     const o = others.map(outsider => outsider.address)
 
     let testToken: TestToken
-    let dataUnionSidechain: DataUnionSidechain
-    let dataUnionSidechainAgent: DataUnionSidechain
-    let dataUnionSidechainMember0: DataUnionSidechain
-    let mockAMB: MockAMB
-    let mockTokenMediator: MockTokenMediator
+    let feeOracle: DefaultFeeOracle
+    let dataUnionSidechain: DataUnionTemplate
+    let dataUnionSidechainAgent: DataUnionTemplate
+    let dataUnionSidechainMember0: DataUnionTemplate
 
     before(async () => {
-        testToken = await deployContract(creator, TestTokenJson, ["name", "symbol"]) as TestToken
-        await testToken.mint(creator.address, parseEther("10000"))
-
-        mockAMB = await deployContract(creator, MockAMBJson, []) as MockAMB
-        mockTokenMediator = await deployContract(creator, MockTokenMediatorJson, [testToken.address, mockAMB.address]) as MockTokenMediator
+        testToken = await deployContract(dao, TestTokenJson, ["name", "symbol"]) as TestToken
+        await testToken.mint(dao.address, parseEther("100000"))
+        feeOracle = await deployContract(dao, FeeOracleJson) as DefaultFeeOracle
+        await feeOracle.initialize(parseEther("0.01"), dao.address)
 
         log("List of relevant addresses:")
-        log("  mockTokenMediator: ", mockTokenMediator.address)
-        log("  mockAMB: ", mockAMB.address)
-        log("  testToken: ", testToken.address)
-        log("  creator: ", creator.address)
+        log("  testToken: %s", testToken.address)
+        log("  dao: %s", dao.address)
+        log("  admin: %s", admin.address)
         log("  agents: %o", a)
         log("  members: %o", m)
-        log("  outsider addresses used in tests: %o", others.map(x => x.address))
+        log("  outsider addresses used in tests: %o", o)
     })
 
     beforeEach(async () => {
-        const dummyAddress = a[0]
-
-        dataUnionSidechain = await deployContract(creator, DataUnionSidechainJson, []) as DataUnionSidechain
+        dataUnionSidechain = await deployContract(admin, DataUnionTemplateJson, []) as DataUnionTemplate
         dataUnionSidechainAgent = dataUnionSidechain.connect(agents[1])
         dataUnionSidechainMember0 = dataUnionSidechain.connect(members[0])
 
         // function initialize(
         //     address initialOwner,
         //     address tokenAddress,
-        //     address tokenMediatorAddress,
         //     address[] memory initialJoinPartAgents,
-        //     address mainnetDataUnionAddress,
         //     uint256 defaultNewMemberEth,
         //     uint256 initialAdminFeeFraction,
-        //     uint256 initialDataUnionFeeFraction,
-        //     address initialDataUnionBeneficiary
+        //     address protocolFeeOracleAddress,
+        //     string calldata metadataJsonString
         // )
         await dataUnionSidechain.initialize(
-            creator.address,
+            admin.address,
             testToken.address,
-            mockTokenMediator.address,
             a,
-            dummyAddress,
             "1",
-            parseEther("0.1"),
-            parseEther("0.1"),
-            duBeneficiary.address
+            parseEther("0.09"), // total fees are 1% + 9% = 10%
+            feeOracle.address,
+            "{}"
         )
         await dataUnionSidechainAgent.addMembers(m)
 
-        log(`DataUnionSidechain initialized at ${dataUnionSidechain.address}`)
+        log(`DataUnionTemplate initialized at ${dataUnionSidechain.address}`)
     })
 
     it("distributes earnings correctly", async () => {
@@ -129,37 +118,44 @@ describe("DataUnionSidechain", () => {
         // repeating it should do nothing (also not throw)
         await dataUnionSidechain.connect(randomOutsider).refreshRevenue()
 
-        expect(await dataUnionSidechain.totalEarnings()).to.equal(2400)
-        expect(await dataUnionSidechain.totalAdminFees()).to.equal(300)
-        expect(await dataUnionSidechain.totalDataUnionFees()).to.equal(300)
-        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(800)
-        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(800)
-        expect(await dataUnionSidechain.getEarnings(m[2])).to.equal(800)
+        expect(await dataUnionSidechain.totalEarnings()).to.equal(2700)
+        expect(await dataUnionSidechain.totalAdminFees()).to.equal(270)
+        expect(await dataUnionSidechain.getEarnings(admin.address)).to.equal(270)
+        expect(await dataUnionSidechain.totalProtocolFees()).to.equal(30)
+        expect(await dataUnionSidechain.getEarnings(dao.address)).to.equal(30)
+        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(m[2])).to.equal(900)
 
         // drop a member, send more tokens, check accounting
         await expect(dataUnionSidechainAgent.partMember(m[0])).to.emit(dataUnionSidechain, "MemberParted")
-        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(800)
+        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(900)
         await testToken.transfer(dataUnionSidechain.address, "2000")
         await dataUnionSidechain.connect(randomOutsider).refreshRevenue()
-        expect(await dataUnionSidechain.totalEarnings()).to.equal(4000)
-        expect(await dataUnionSidechain.totalAdminFees()).to.equal(500)
-        expect(await dataUnionSidechain.totalDataUnionFees()).to.equal(500)
-        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(800)
-        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(1600)
-        expect(await dataUnionSidechain.getEarnings(m[2])).to.equal(1600)
+        expect(await dataUnionSidechain.totalEarnings()).to.equal(4500)
+        expect(await dataUnionSidechain.totalAdminFees()).to.equal(450)
+        expect(await dataUnionSidechain.getEarnings(admin.address)).to.equal(450)
+        expect(await dataUnionSidechain.totalProtocolFees()).to.equal(50)
+        expect(await dataUnionSidechain.getEarnings(dao.address)).to.equal(50)
+        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(1800)
+        expect(await dataUnionSidechain.getEarnings(m[2])).to.equal(1800)
         await expect(dataUnionSidechainAgent.addMember(m[0])).to.emit(dataUnionSidechain, "MemberJoined")
 
         // add a member, send tokens, check accounting
         await expect(dataUnionSidechainAgent.addMember(newMember.address)).to.emit(dataUnionSidechain, "MemberJoined")
         await testToken.transfer(dataUnionSidechain.address, "4000")
         await dataUnionSidechain.connect(randomOutsider).refreshRevenue()
-        expect(await dataUnionSidechain.totalEarnings()).to.equal(7200)
-        expect(await dataUnionSidechain.totalAdminFees()).to.equal(900)
-        expect(await dataUnionSidechain.totalDataUnionFees()).to.equal(900)
-        expect(await dataUnionSidechain.getEarnings(newMember.address)).to.equal(800)
-        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(1600)
-        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(2400)
-        expect(await dataUnionSidechain.getEarnings(m[2])).to.equal(2400)
+        expect(await dataUnionSidechain.totalEarnings()).to.equal(8100)
+        expect(await dataUnionSidechain.totalAdminFees()).to.equal(810)
+        expect(await dataUnionSidechain.getEarnings(admin.address)).to.equal(810)
+        expect(await dataUnionSidechain.totalProtocolFees()).to.equal(90)
+        expect(await dataUnionSidechain.getEarnings(dao.address)).to.equal(90)
+        expect(await dataUnionSidechain.getEarnings(newMember.address)).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(1800)
+        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(2700)
+        expect(await dataUnionSidechain.getEarnings(m[2])).to.equal(2700)
         await expect(dataUnionSidechainAgent.partMember(newMember.address)).to.emit(dataUnionSidechain, "MemberParted")
     })
 
@@ -216,15 +212,17 @@ describe("DataUnionSidechain", () => {
     it("getEarnings", async () => {
         await expect(dataUnionSidechain.getEarnings(o[0])).to.be.revertedWith("error_notMember")
         await expect(dataUnionSidechain.getEarnings(a[0])).to.be.revertedWith("error_notMember")
-        await expect(dataUnionSidechain.getEarnings(creator.address)).to.be.revertedWith("error_notMember")
+        await expect(dataUnionSidechain.getEarnings(admin.address)).to.be.revertedWith("error_notMember")
         expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(0)
 
         await testToken.transfer(dataUnionSidechain.address, "3000")
         await dataUnionSidechain.refreshRevenue()
 
-        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(800)
-        expect(await dataUnionSidechain.getEarnings(creator.address)).to.equal(300)
-        expect(await dataUnionSidechain.getEarnings(duBeneficiary.address)).to.equal(300)
+        expect(await dataUnionSidechain.getEarnings(m[0])).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(m[1])).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(m[2])).to.equal(900)
+        expect(await dataUnionSidechain.getEarnings(admin.address)).to.equal(270)
+        expect(await dataUnionSidechain.getEarnings(dao.address)).to.equal(30)
     })
 
     async function getBalances(addresses: EthereumAddress[]) {
@@ -242,7 +240,7 @@ describe("DataUnionSidechain", () => {
         await testToken.transfer(dataUnionSidechain.address, "3000")
         await dataUnionSidechain.refreshRevenue()
         await expect(dataUnionSidechain.withdrawMembers(m, false)).to.emit(dataUnionSidechain, "EarningsWithdrawn")
-        expect(await getBalanceIncrements(m, balances)).to.deep.equal([ 800, 800, 800 ])
+        expect(await getBalanceIncrements(m, balances)).to.deep.equal([ 900, 900, 900 ])
     })
 
     it("withdrawAll", async () => {
@@ -253,7 +251,7 @@ describe("DataUnionSidechain", () => {
         await expect(dataUnionSidechainMember0.withdrawAll(m[0], false)).to.emit(dataUnionSidechain, "EarningsWithdrawn")
         await expect(dataUnionSidechain.withdrawAll(m[1], false)).to.emit(dataUnionSidechain, "EarningsWithdrawn")
         await dataUnionSidechain.withdrawAll(m[1], false)    // this should do nothing, also not revert
-        expect(await getBalanceIncrements(m, balances)).to.deep.equal([ 800, 800, 0 ])
+        expect(await getBalanceIncrements(m, balances)).to.deep.equal([ 900, 900, 0 ])
     })
 
     it("withdrawAllTo", async () => {
@@ -265,7 +263,7 @@ describe("DataUnionSidechain", () => {
         const after = await testToken.balanceOf(o[0])
 
         const diff = after.sub(before)
-        expect(diff).to.equal(800)
+        expect(diff).to.equal(900)
     })
 
     it("withdrawToSigned", async () => {
@@ -302,13 +300,13 @@ describe("DataUnionSidechain", () => {
         await expect(dataUnionSidechainRecipient.withdrawAllToSigned(m[0], r, false, signature)).to.be.revertedWith("error_badSignature")
         await expect(dataUnionSidechainRecipient.withdrawAllToSigned(m[1], r, false, signature)).to.emit(dataUnionSidechain, "EarningsWithdrawn")
 
-        expect(await testToken.balanceOf(r)).to.equal(800)
+        expect(await testToken.balanceOf(r)).to.equal(900)
     })
 
     it("transferToMemberInContract", async () => {
         await testToken.approve(dataUnionSidechain.address, "2000")
-        await dataUnionSidechain.transferToMemberInContract(o[0], "1000")
-        await dataUnionSidechain.transferToMemberInContract(m[0], "1000")
+        await dataUnionSidechain.connect(dao).transferToMemberInContract(o[0], "1000")
+        await dataUnionSidechain.connect(dao).transferToMemberInContract(m[0], "1000")
         expect(await dataUnionSidechain.getWithdrawableEarnings(o[0])).to.equal(1000)
         expect(await dataUnionSidechain.getWithdrawableEarnings(m[0])).to.equal(1000)
 
@@ -319,6 +317,13 @@ describe("DataUnionSidechain", () => {
         await expect(dataUnionSidechain.transferToMemberInContract(m[0], "777")).to.be.revertedWith("error_transfer")
     })
 
+    it("transferToMemberInContract using ERC677", async () => {
+        await testToken.transferAndCall(dataUnionSidechain.address, "1000", o[0])
+        await testToken.transferAndCall(dataUnionSidechain.address, "1000", m[0])
+        expect(await dataUnionSidechain.getWithdrawableEarnings(o[0])).to.equal(1000)
+        expect(await dataUnionSidechain.getWithdrawableEarnings(m[0])).to.equal(1000)
+    })
+
     it("transferWithinContract", async () => {
         assert(await testToken.transfer(dataUnionSidechain.address, "3000"))
         await dataUnionSidechain.refreshRevenue()
@@ -327,27 +332,27 @@ describe("DataUnionSidechain", () => {
         // await expect(dataUnionSidechain.transferWithinContract(m[1], "100")).to.be.revertedWith("error_notMember")
         await expect(dataUnionSidechainMember0.transferWithinContract(m[1], "100")).to.emit(dataUnionSidechain, "TransferWithinContract")
         await expect(dataUnionSidechainMember0.transferWithinContract(o[1], "100")).to.emit(dataUnionSidechain, "TransferWithinContract")
-        expect(await dataUnionSidechain.getWithdrawableEarnings(m[0])).to.equal(600)
-        expect(await dataUnionSidechain.getWithdrawableEarnings(m[1])).to.equal(900)
-        expect(await dataUnionSidechain.getWithdrawableEarnings(m[2])).to.equal(800)
+        expect(await dataUnionSidechain.getWithdrawableEarnings(m[0])).to.equal(700)  // = 900 - 100 - 100
+        expect(await dataUnionSidechain.getWithdrawableEarnings(m[1])).to.equal(1000) // = 900 + 100
+        expect(await dataUnionSidechain.getWithdrawableEarnings(m[2])).to.equal(900)  // no changes
         expect(await dataUnionSidechain.getWithdrawableEarnings(o[1])).to.equal(100)
         // those who received some in-contract balance but aren't members should be marked inactive by _increaseBalance
         expect(await dataUnionSidechain.inactiveMemberCount()).to.equal(3)
         expect((await dataUnionSidechain.memberData(o[1])).status).to.equal(2)
-        expect((await dataUnionSidechain.memberData(duBeneficiary.address)).status).to.equal(2)
-        expect((await dataUnionSidechain.memberData(creator.address)).status).to.equal(2)
+        expect((await dataUnionSidechain.memberData(dao.address)).status).to.equal(2)
+        expect((await dataUnionSidechain.memberData(admin.address)).status).to.equal(2)
     })
 
     it("getStats", async () => {
         // test send with transferAndCall. refreshRevenue not needed in this case
         await testToken.transferAndCall(dataUnionSidechain.address, "3000", "0x")
 
-        await dataUnionSidechainMember0.withdraw(m[0], "500", true)
+        await dataUnionSidechainMember0.withdraw(m[0], "500", false)
         const [
             totalRevenue,
             totalEarnings,
             totalAdminFees,
-            totalDataUnionFees,
+            totalProtocolFees,
             totalEarningsWithdrawn,
             activeMemberCount,
             inactiveMemberCount,
@@ -355,26 +360,22 @@ describe("DataUnionSidechain", () => {
             joinPartAgentCount
         ] = await dataUnionSidechain.getStats()
         expect(totalRevenue).to.equal(3000)
-        expect(totalEarnings).to.equal(2400)
-        expect(totalAdminFees).to.equal(300)
-        expect(totalDataUnionFees).to.equal(300)
+        expect(totalEarnings).to.equal(2700)
+        expect(totalAdminFees).to.equal(270)
+        expect(totalProtocolFees).to.equal(30)
         expect(totalEarningsWithdrawn).to.equal(500)
         expect(activeMemberCount).to.equal(3)
-        expect(inactiveMemberCount).to.equal(0) // creator and duBeneficiary are cleaned out of this number though they show up in the "inactiveMemberCount"
-        expect(lifetimeMemberEarnings).to.equal(800)
+        expect(inactiveMemberCount).to.equal(0) // admin and dao are cleaned out of this number though they show up in the "inactiveMemberCount"
+        expect(lifetimeMemberEarnings).to.equal(900)
         expect(joinPartAgentCount).to.equal(2)
     })
 
-    // Of course there is no "withdraw to mainnet" in test.
-    // Instead what happens in DataUnionSidechain is a call to TokenMediator
-    it("withdraw to mainnet", async () => {
+    // withdraw to mainnet is deprecated
+    it("fails calls to withdraw to mainnet", async () => {
         await testToken.transfer(dataUnionSidechain.address, "3000")
-        // for coverage completeness: also exercise the bridge's token transfer callback. Simply calls refreshRevenue.
-        await dataUnionSidechain.onTokenBridged(testToken.address, "3000", "0x")
-        await expect(dataUnionSidechainMember0.withdraw(m[0], "100", true)).to.emit(dataUnionSidechain, "EarningsWithdrawn")
 
         // TestToken blocks transfers with this magic amount
-        await expect(dataUnionSidechainMember0.withdraw(m[0], "666", true)).to.be.revertedWith("error_transfer")
+        await expect(dataUnionSidechainMember0.withdraw(m[0], "100", true)).to.be.revertedWith("error_sendToMainnetDeprecated")
     })
 
     it("fails to withdraw more than earnings", async () => {
@@ -388,17 +389,14 @@ describe("DataUnionSidechain", () => {
 
     it("fails to initialize twice", async () => {
         const a = agents.map(agent => agent.address)
-        const dummyAddress = a[0]
         await expect(dataUnionSidechain.initialize(
-            creator.address,
+            admin.address,
             testToken.address,
-            mockTokenMediator.address,
             a,
-            dummyAddress,
             "1",
             parseEther("0.1"),
-            parseEther("0.1"),
-            duBeneficiary.address
+            feeOracle.address,
+            "{}"
         )).to.be.revertedWith("error_alreadyInitialized")
     })
 
@@ -430,25 +428,56 @@ describe("DataUnionSidechain", () => {
         expect(await dataUnionSidechain.owner()).to.equal(o[0])
 
         await expect(dataUnionSidechain.transferOwnership(o[0])).to.be.revertedWith("error_onlyOwner")
-        await dataUnionSidechain.connect(others[0]).transferOwnership(creator.address)
+        await dataUnionSidechain.connect(others[0]).transferOwnership(admin.address)
         await expect(dataUnionSidechain.claimOwnership()).to.emit(dataUnionSidechain, "OwnershipTransferred")
-        expect(await dataUnionSidechain.owner()).to.equal(creator.address)
+        expect(await dataUnionSidechain.owner()).to.equal(admin.address)
     })
 
     it("rejects unexpected ERC677 tokens", async () => {
-        const randomToken = await deployContract(creator, TestTokenJson, ["random", "RND"]) as TestToken
-        await randomToken.mint(creator.address, parseEther("10000"))
+        const randomToken = await deployContract(admin, TestTokenJson, ["random", "RND"]) as TestToken
+        await randomToken.mint(admin.address, parseEther("10000"))
         await expect(randomToken.transferAndCall(dataUnionSidechain.address, "1000", "0x")).to.be.revertedWith("error_onlyTokenContract")
     })
 
-    it("rejects fees that sum above 1.0", async () => {
-        await expect(dataUnionSidechain.setFees(parseEther("0.5"), parseEther("0.6"))).to.be.revertedWith("error_fees")
+    it("rejects admin fee that would cause total fees sum above 1.0", async () => {
+        await expect(dataUnionSidechain.setAdminFee(parseEther("0.995"))).to.be.revertedWith("error_adminFee")
+    })
+
+    it("adjusts an admin fee that would cause total fees sum above 1.0", async () => {
+        await expect(dataUnionSidechain.setAdminFee(parseEther("0.9"))).to.emit(dataUnionSidechain, "AdminFeeChanged")
+        expect(await dataUnionSidechain.adminFeeFraction()).to.equal(parseEther("0.9"))
+        await feeOracle.setFee(parseEther("0.2"))
+        assert(await testToken.transfer(dataUnionSidechain.address, "3000"))
+        await dataUnionSidechain.refreshRevenue()
+        expect(await dataUnionSidechain.adminFeeFraction()).to.equal(parseEther("0.8"))
+    })
+
+    it("lets only admin change the metadata", async () => {
+        await expect(dataUnionSidechain.connect(members[0]).setMetadata("foo")).to.be.revertedWith("error_onlyOwner")
+        expect(await dataUnionSidechain.metadataJsonString()).to.equal("{}")
+        await expect(dataUnionSidechain.connect(admin).setMetadata("foo")).to.emit(dataUnionSidechain, "MetadataChanged")
+        expect(await dataUnionSidechain.metadataJsonString()).to.equal("foo")
+    })
+
+    it("lets only admin change the admin fee", async () => {
+        await expect(dataUnionSidechain.connect(members[0]).setAdminFee(parseEther("0.5"))).to.be.revertedWith("error_onlyOwner")
+        expect(await dataUnionSidechain.adminFeeFraction()).to.equal(parseEther("0.09"))
+        await expect(dataUnionSidechain.connect(admin).setAdminFee(parseEther("0.5"))).to.emit(dataUnionSidechain, "AdminFeeChanged")
+        expect(await dataUnionSidechain.adminFeeFraction()).to.equal(parseEther("0.5"))
     })
 
     it("cannot swap modules after locking", async () => {
         const dummyAddress = "0x1234567890123456789012345678901234567890"
         await expect(dataUnionSidechain.setWithdrawModule(dummyAddress)).to.emit(dataUnionSidechain, "WithdrawModuleChanged")
+        await expect(dataUnionSidechain.addJoinListener(dummyAddress)).to.emit(dataUnionSidechain, "JoinListenerAdded")
+        await expect(dataUnionSidechain.addPartListener(dummyAddress)).to.emit(dataUnionSidechain, "PartListenerAdded")
+        await expect(dataUnionSidechain.removeJoinListener(dummyAddress)).to.emit(dataUnionSidechain, "JoinListenerRemoved")
+        await expect(dataUnionSidechain.removePartListener(dummyAddress)).to.emit(dataUnionSidechain, "PartListenerRemoved")
         await dataUnionSidechain.lockModules()
         await expect(dataUnionSidechain.setWithdrawModule(dummyAddress)).to.be.revertedWith("error_modulesLocked")
+        await expect(dataUnionSidechain.addJoinListener(dummyAddress)).to.be.revertedWith("error_modulesLocked")
+        await expect(dataUnionSidechain.addPartListener(dummyAddress)).to.be.revertedWith("error_modulesLocked")
+        await expect(dataUnionSidechain.removeJoinListener(dummyAddress)).to.be.revertedWith("error_modulesLocked")
+        await expect(dataUnionSidechain.removePartListener(dummyAddress)).to.be.revertedWith("error_modulesLocked")
     })
 })
